app: "Canvas Size" dialog takes too much...
vertical space.
Since the changes to add template ability, there are just too
many widgets stacked up vertically. Let's use the horizontal space
a little better.

To do this, Jehan moved the preview on the right side of the Canvas Size
and Offset number fields, and the "Center" button just below it.