build/windows: Port Installer script to .ps1
This commit unites two scripts in one, which is more efficient to maintain.
PowerShell was choosed because we don't need MSYS2 to distribute with Inno
and PS have a good number of cmdlets, unlike the ancient CMD.

Also, meson.build was changed changed to avoid GTK hardcoding in the future.