gitlab-ci: Reduce redundancy on Windows cross and native jobs
This commit created definitions to avoid repeating the same CI code
Some people (like the GitLab devs) prefer this to be done in multiple .yml's
using 'worfklow:' but this pratice, for me, makes the CI overly complicated
and very slow to read. So, let's just use simple defs in one readable file.

The manual native arch pipelines were improved a bit. Now they always generate
a .zip package to ensure that everything is in place and facilitate testing.