gitlab-ci: Move 'analysis' stage to a more "mainstream" position
I suppose the intention of having this stage at the end was to make clear that
no job depends on it. However, this place is not so common. The appropriate
and usual place is after building, even if this isn't dynamic analysis (tests).

The new place is a bit more clear to devs used at "build, test, deploy" praxis
and maybe will be now better noticed by the existing devs before distributing.
The stage continues to be non-mandatory to pass, but clarity is always welcome.

---

Also, dropped the deprecated "only:" (in clang-format job) in favor of "rules:"
and maked the artifacts paths of clang-format and cppcheck more consistent.