gitlab-ci, build: Use 'bundling' naming instead of 'packaging'
We now use of the word 'bundle' to signify "program files in the same prefix"
(e.g. .flatpak, .zip, .app). This is in line with our source and dev-docs.

The word 'package' normally means "program files distributed for install in
the same prefix or not" (e.g. .flatpakref, .msix, .dmg) and will not be used.
This is in line with CMake docs, but meson prefers to call 'dist' instead.