gitlab-ci: Unify artifacts "expire_in" time
The 'dependencies' stage continue to expire in 2 hours. Now, 'gimp'
expires in 2 days (to avoid time zone limbo), 'analysis' in 4 (to
have a moderate time to analising) and 'dist' in 8 days (to have a
+1 day artifact if something goes wrong in the weekly schedule day).

This can potentially boost the storage usage again... or not. I've
been enabling things like auto cancelling of jobs to prevent this.