README: Port to Markdown
For developers, this format is easier to read than the plain text due to
syntax marking on IDEs. For first time contributors, it is easy to read on
smarphone since it is rendered with responsiveness taked into account by
the "HTML" preview. Also, the HTML preview makes esier click on links.

These above are objective, undisputed and unrefutable features of MD.