gitlab-ci: Fix custom builds being interrupted
This bug was introduced with c064148a and it's similar to fb7a9954. Now, that
we are using 'interruptible: false' in the custom rule, this is probably fixed.

Also, organize a bit better some pipelines with official GitLab .yml wizardy.
I was hesitating to use "<<: *" since this creates a new layer of complexity,
but this was the only way and the trick have a nice glowing mark by the way.