build/windows: Fix Inno wrong components size calc (part 1)
This solves an old bug that makes Inno report GIMP as being almost 2GB,
which is not true (it's 1.2GB at worst). That was probably caused by
the use of components intersections (e.g.: gs + gimp, py + deps), which
Inno does not handle well (sometimes it not even displayed the size).

TWAIN component is still reporting a wrong size since it's a totally
different beast, it will be fixed later.