build/windows: Prepare MSIX versioning to RC1 and beyond
This fixes the pseudo-revision trick when we have a zeroed micro version,
which is the case of the release candidates and first stable version.

First, make the script versioning system able to detect release candidates.
Otherwise, we willn't be able to use the "GIMP.GIMPPreview" identity (which
was previously tied to GIMP_UNSTABLE var, which is not set to 1 on RCs.

---

This also fixes the pseudo-revision trick when we have a zeroed micro version,
which is the case of the release candidates and the first stable version.
The versioning now have a different aproach from COMMIT_SHA.