operations: Improve speed of HSV Legacy Layer mode
Previously, we used babl_process () multiple times for each pixel
when painting with HSV legacy layer modes. This caused a great
deal of slow down when painting.

This patch moves the babl_process () code before the loop and
processes the entire sample all at once. This seems to speed up
the operation significantly.